sys/riotboot_usb_dfu: make feature no_idle_thread optional
`no_idle_thread` was introduced to save some RAM and ROM because there is no idle thread is used in that case. However, insisting that on feature the use of `riotboot_dfu` on platforms that doesn't provide that feature even if saving RAM and ROM doesn't matter.
Therefore the dependency of `riotboot_dfu` on this feature is change to optional. If the platform provides, the feature is used and a idle thread is not created.