boards/common/gd32v: cleanup default I2C and SPI config
The macros `I2C_DEV_1_USED` and `SPI_DEV_1_USED` are now used with the values 0 and 1 to be able enable the use of the respective second interface but also to be able to disable them if the board definition enables them by default.

The second I2C device `I2C_DEV(1)` and second SPI device `SPI_DEV(1)` are now disabled by default.

The second SPI device `SPI_DEV(1)` now uses PB5 as default CS signal instead of PA4 to keep PA4 free for ADC or DAC even if `SPI_DEV(1)` is used, for example for the TFT display.